Update Readme.md
Added instructions for compilation. We are using Python 3.3, right?!?
Added licensing reference.
Added Q&A section for questions regarding the program. (We don't have to add anything right now. It might be a good idea to have if people start asking why we are using Python x.y, ect. ect...)